requirements.txt
asgiref==3.6.0
dj-database-url==1.2.0
Django==3.2.7
django-admin-interface==0.24.2
django-ckeditor==6.5.1
django-cleanup==6.0.0
django-colorfield==0.8.0
django-js-asset==2.0.0
gunicorn==20.1.0
Pillow==9.4.0
psycopg2==2.9.5
python-decouple==3.7
pytz==2022.7.1
sqlparse==0.4.3
whitenoise==6.3.0
